changed from server side now using cdn
