+++ R-REC-NEW/dashboard.css
@@ -1,0 +1,130 @@
+
+/* navigation start*/
+.container{
+    width: 90%;
+    margin: 0 auto;
+  }
+  .nav{
+    display: flex;
+    box-shadow: 0 0.1rem 0.2rem rgba(0, 0, 0, 0.2);
+    background-color: white;
+    padding: 0 1rem;
+  }
+  #menuButton{
+    display: none;
+  }
+  ul{
+    justify-content:end;
+    display: flex;
+    list-style: none;
+  }
+  ul li{
+    margin: 0;
+    padding: 0.5rem 0;
+    margin-left: 1.5rem;
+    color: #ff6d05;
+  }
+  .team{
+    background: #F5F5F5;
+    border-radius: 9px;
+    padding: 0.5rem 1rem;
+  }
+  .global{
+    background: #e3823c;
+    border-radius: 9px;
+    padding: 0.5rem 1rem; 
+    color: white;
+  }
+  .global i {
+    background-color: white;
+    padding: 0.2rem 0.2rem;
+    border-radius: 50%;
+    margin-right: 0.6rem;
+    color:#FF9B53 ;
+  }
+  .global i .g{
+    padding: 0 0 0 0;
+    margin: 2rem 0.1rem;
+    font-weight: bolder;
+    text-align: center;
+  }
+  img {
+    width: 200px; /* Set the desired width */
+    height: auto; 
+    padding: 0.5rem 0;
+    
+  }
+/* navigation end */
+
+.cart-container {
+    background-color: #BCBCBC;
+    width: 100%;
+    margin: 0;
+}
+
+.cart{
+    display: flex;
+    justify-content: space-between;
+    margin: 0 4rem;
+    align-items: center;
+}
+span .text{
+    color: white;
+}
+.columns{
+    display: flex;
+    margin: 0 4rem;
+    gap: 2rem;
+}
+.category{
+    background-color: white;
+    margin-top: 2rem;
+    padding: 0.5rem 2rem;
+    box-shadow: 0 0.1rem 0.2rem rgba(0, 0, 0, 0.2);
+    border-radius: 0.5rem;
+}
+.category-heading {
+    position: relative;
+    display: inline-block;
+    margin-bottom: 0.8rem; /* Adjust as needed */
+    width: 100%;
+}
+
+.category-heading::after {
+    content: '';
+    position: absolute;
+    width: 100%;
+    height: 0.1rem; /* Adjust as needed */
+    background-color: rgb(206, 201, 201); /* Replace with your desired color */
+    bottom: 0;
+    left: 0;
+}
+
+.category-section{
+    margin-bottom: 10px; /* Adjust the margin */
+    display: flex;
+    align-items: center;
+    justify-content: space-between;
+}
+.category-section label{
+    font-weight: 600;
+    font-size: 0.85rem;
+}
+.cards{
+    width: 100%;
+}
+.cards .card-row {
+    display: flex;
+    gap:2rem;
+    margin-bottom: 2rem; /* Adjust as needed */
+ }
+
+.card {
+    width: 30%;
+    border: 1px solid #ccc; /* Add styling as needed */
+    padding: 20px; /* Adjust as needed */
+    background-color: white;
+    border-radius: 0.5rem;
+    margin-top: 2rem;
+    
+}
