+++ R-REC-NEW/dashboard.html
@@ -1,0 +1,152 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+    <meta charset="UTF-8">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <title>Dashboard</title>
+    <link href="https://fonts.googleapis.com/css2?family=Roboto:wght@100;300;500;700;900&display=swap" rel="stylesheet">
+    <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/6.0.0-beta3/css/all.min.css">
+    <link rel="stylesheet" href="dashboard.css">
+    <!-- <link rel="stylesheet" href="index.css"> -->
+
+</head>
+<body>
+    <header>
+        <nav class="nav">
+            <div class="logo">
+                 <a href="/">
+                 <img src="images/renewvia.webp" alt="renewvia-logo">
+                 </a>
+                <a onclick="toggleMenu()" id="menuButton"><i class="fas fa-bars"></i></a>
+            </div>
+            <div class="hide container " id="info">
+                <ul>
+                    <router-link to="/about">
+                    <li>ABOUT US</li>
+                    </router-link>
+                    <li>WHAT ARE R-RECs</li>
+                    <li>PRODUCTS</li>
+                    <li class="team">OUR TEAM</li>
+                    <li class="global">
+                        <i class="fas fa-g"><span class="g"></span></i>globalcorp</li>
+                </ul>
+            </div>        
+        </nav>
+        </header>
+        <div class="cart-container">
+            <span class="cart">
+                <h1 class="text">
+                    Get started by choosing your R-RECs below
+                </h1>
+                <svg xmlns="http://www.w3.org/2000/svg" width="30" height="30" fill="white" class="bi bi-cart3" viewBox="0 0 16 16">
+                <path d="M0 1.5A.5.5 0 0 1 .5 1H2a.5.5 0 0 1 .485.379L2.89 3H14.5a.5.5 0 0 1 .49.598l-1 5a.5.5 0 0 1-.465.401l-9.397.472L4.415 11H13a.5.5 0 0 1 0 1H4a.5.5 0 0 1-.491-.408L2.01 3.607 1.61 2H.5a.5.5 0 0 1-.5-.5zM3.102 4l.84 4.479 9.144-.459L13.89 4H3.102zM5 12a2 2 0 1 0 0 4 2 2 0 0 0 0-4zm7 0a2 2 0 1 0 0 4 2 2 0 0 0 0-4zm-7 1a1 1 0 1 1 0 2 1 1 0 0 1 0-2zm7 0a1 1 0 1 1 0 2 1 1 0 0 1 0-2z"/>
+                </svg>
+            </span>
+        </div>
+        <div class="columns">
+            <div class="category">
+                <h2 class="category-heading">Category</h2>
+                <div class="category-section">
+                    <label for="section1">All</label>
+                    <input type="checkbox" id="section1" name="section1">
+                </div>
+    
+                <div class="category-section">
+                    <label for="section2">Solar</label>
+                    <input type="checkbox" id="section2" name="section2">
+                </div>
+    
+                <div class="category-section">
+                    <label for="section3">Hydro</label>
+                    <input type="checkbox" id="section3" name="section3">
+                </div>
+    
+                <div class="category-section">
+                    <label for="section4">Wind</label>
+                    <input type="checkbox" id="section4" name="section4">
+                </div>
+    
+                <div class="category-section">
+                    <label for="section5">Geothermal</label>
+                    <input type="checkbox" id="section5" name="section5">
+                </div>
+    
+                <div class="category-section">
+                    <label for="section6">Biomass</label>
+                    <input type="checkbox" id="section6" name="section6">
+                </div>
+    
+                <h2 class="category-heading">Country</h2>
+                <div class="category-section">
+                    <label for="section1">All</label>
+                    <input type="checkbox" id="section1" name="section1">
+                </div>
+    
+                <div class="category-section">
+                    <label for="section2">Alabama</label>
+                    <input type="checkbox" id="section2" name="section2">
+                </div>
+    
+                <div class="category-section">
+                    <label for="section3">Colorado</label>
+                    <input type="checkbox" id="section3" name="section3">
+                </div>
+    
+                <div class="category-section">
+                    <label for="section4">Georgia</label>
+                    <input type="checkbox" id="section4" name="section4">
+                </div>
+    
+                <div class="category-section">
+                    <label for="section5">North Carolina</label>
+                    <input type="checkbox" id="section5" name="section5">
+                </div>
+    
+                <div class="category-section">
+                    <label for="section6">Texas</label>
+                    <input type="checkbox" id="section6" name="section6">
+                </div>
+    
+                <div class="category-section">
+                    <label for="section7">Virginia</label>
+                    <input type="checkbox" id="section7" name="section7">
+                </div>
+    
+                <div class="category-section">
+                    <label for="section8">Zambia</label>
+                    <input type="checkbox" id="section8" name="section8">
+                </div>
+            </div>
+            <div class="cards">
+                <div class="card-row">
+                    <div class="card">
+                        <div class="section1">
+                            <h3>North Carolina R-REC</h3>
+                            <span>$35.00/REC</span> 
+                            <p>A blend of RECs from our diverse portfolio of renewable energy projects. 1 REC = 1MWh (1000 kWh)</p>
+                        </div>
+                        <div class="section2">
+                            <h2>$70.00</h2>
+                            <p>
+                                <button>A</button>
+                                <button>A</button>
+                            </p>
+                        </div>
+                    </div>
+                    <div class="card">Card 2</div>
+                    <div class="card">Card 3</div>
+                </div>
+    
+                <div class="card-row">
+                    <div class="card">Card 1</div>
+                    <div class="card">Card 2</div>
+                    <div class="card">Card 3</div>
+                </div>
+            </div>
+        </div>
+    </header>
+
+
+    <!-- <script src="https://unpkg.com/vue@3/dist/vue.global.js"></script> -->
+</body>
+</html>
