--- R-REC-NEW/package.json
@@ -1,19 +1,0 @@
-{
-  "name": "rrecs-app",
-  "version": "0.1.0",
-  "private": true,
-  "scripts": {
-    "serve": "vue-cli-service serve",
-    "build": "vue-cli-service build"
-  },
-  "dependencies": {
-    "chart.js": "^4.4.0",
-    "core-js": "^3.8.3",
-    "vue": "^3.2.13",
-    "vue-chartjs": "^5.2.0"
-  },
-  "devDependencies": {
-    "@vue/cli-plugin-babel": "~5.0.0",
-    "@vue/cli-service": "~5.0.0"
-  }
-}
