--- R-REC-NEW/src/components/Index.vue
@@ -1,219 +1,0 @@
-<template>
-  <!DOCTYPE html>
-    <html lang="en">
-        <head>
-            <meta charset="utf-8">
-            <meta http-equiv="X-UA-Compatible" content="IE=edge">
-            <title>R-RECs App</title>
-            <meta name="description" content="">
-            <meta name="viewport" content="width=device-width, initial-scale=1">
-            <link rel="preconnect" href="https://fonts.googleapis.com">
-            <link rel="preconnect" href="https://fonts.gstatic.com" crossorigin>
-            <link href="https://fonts.googleapis.com/css2?family=Roboto:wght@100;300;500;700;900&display=swap" rel="stylesheet">
-            <!-- <link rel="stylesheet" href="index.css"> -->
-            <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/6.0.0-beta3/css/all.min.css">
-            
-        </head>
-        <body>
-            <!-- section nav -->
-            <header>
-                <Navigation></Navigation>
-            </header>
-            <!-- end section nav -->
-
-            <!-- section 2  -->
-            <div>
-                <div class="logo1">
-                    <div class="logo-container">
-                        <img src="@/assets/images/Artboard.png" alt="company-logo">
-                    </div>
-                    <div class="column">
-                        <div class="inc">Global Corp, Inc.</div>
-                        <div class="column1">
-                            <div class="binary">0x6Ba...147d</div>
-                            <div class="date">Joined Sep 18 2023</div>
-                        </div>
-                    </div>
-        
-                </div>
-            </div>
-            <!-- end section 2 -->
-            
-            <!-- impact section -->
-            <div class="impact">
-                <div class="my-impact">
-                    <div class="offset">
-                        <div class="impact-icon">
-                            <h1>Total amount of carbon offset</h1>
-                            <i class="fas fa-leaf"></i>
-                        </div>
-                        
-                        <p>{{ carbonAssetsData.gas }} </p>
-                    </div>
-                    <div>
-                        <div class="impact-icon">
-                            <h1>Total amount of renewable energy generation</h1>
-                            <i class="fas fa-solar-panel"></i>
-                        </div>
-                        <p>1906 MWh</p>
-                    </div>
-                </div>
-            </div>
-            <!-- end impact section -->
-
-            <!-- tables -->
-            <div class="tables">
-                <div class="two-tables">
-                    <div class="table-column">
-                        <div class="table1" v-if="carbonAssetsData && carbonAssetsData.length > 0" >
-                            <h1>Carbon Assets</h1>
-                            <table>
-                                <tbody>
-                                    <tr class="assets" v-for="asset in carbonAssetsData" :key="asset.id">
-                                        <td>{{ asset.name }}</td>
-                                        <td><h1>{{ asset.qty }}</h1></td>
-                                        <details>
-                                            <summary></summary>
-                                            <ul v-for="transaction in asset.transactions" :key="transaction.contractAddress" >
-                                                <li>Token Name: {{ transaction.tokenName}}</li>
-                                                <li>Block Number: {{ transaction.blockNumber}}</li>
-                                                <li>Tx Index: {{ transaction.transactionIndex}}</li>
-                                                <li>Timestamp: {{ transaction.timeStamp}}</li>
-                                            </ul>
-                                        </details>
-                                    </tr>
-                                </tbody>
-                                </table>
-                        </div>
-                        <div class="table2" v-if="carbonAssetsData && carbonAssetsData.length == 1">
-                            <h1>Activity</h1>
-                            <table>
-                                <thead>
-                                    <tr>
-                                        <th>Date</th>
-                                        <th>Carbon Assets</th>
-                                        <th>Action</th>
-                                        <th>Amount</th>
-                                        <th>Status</th>
-                                    </tr>
-                                </thead>
-                    
-                                <tbody>
-                                    <tr class="assets" v-for="transaction in carbonAssetsData[0].transactions" :key="transaction.contractAddress">
-                                        <td>{{ transaction.timeStamp}}</td>
-                                        <td> {{ transaction.tokenSymbol}}</td>
-                                        <td>Generation</td>
-                                        <td>{{ transaction.gasUsed}}</td>
-                                        <td><button>Verified</button></td>
-                                    </tr>
-                                </tbody>
-
-                            </table>
-                        </div> 
-                    </div>
-                
-
-                    <!-- chart -->
-                        <div class="pie-chart" v-if="carbonAssetsData && carbonAssetsData.length > 0" >
-                            <div>
-                                <h1>Carbon assets in four states.</h1>
-                                <!-- <div class="chart"></div> -->
-                                <!-- chart -->
-                                <div>
-                                    <canvas ref="myChart"></canvas>
-                                </div>
-
-                                <ul class="key assets">
-                                <li v-for="asset in carbonAssetsData" :key="asset.id">
-                                    <strong class="percent gray">{{ calculatePercentage(asset.qty) }}%</strong>
-                                    <span class="choice">{{ asset.name }}</span>
-                                </li>
-                                </ul> 
-        
-                                <!-- retire & purchase -->
-                                <div class="retire-purchase">
-                                        <div class="retire">Retire Current Credits</div>
-                                        <div class="credits">Mark Credits for sale</div>
-                                        <div class="purchase">Purchase New Credits</div>
-                                    
-                                </div>
-                                <!-- end retire & purchase -->
-                            </div>
-                        </div>    
-                        <!-- end of chart -->
-                </div>
-
-            </div>
-            <!-- end of tables -->
-        </body>
-    </html>
-</template>
-
-<script>
-
-import Chart from 'chart.js/auto';
-
-export default {
-    props: ['carbonAssetsData'],
-    data() {
-        return {
-            chart: null,
-            };
-    },
-    
-    mounted() {
-        this.renderChart();
-    },
-    
-    beforeDestroy() {
-        // Destroy the existing chart instance
-        if (this.chart) {
-            this.chart.destroy();
-        }
-    },
-    
-    watch: {
-        carbonAssetsData: {
-            handler(newData) {
-                // Check if the chart instance exists and destroy it
-                if (this.chart) {
-                    this.chart.destroy();
-                }
-                // Render the chart with the new data
-                this.renderChart();
-            },
-            
-            deep: true, // Watch for changes in nested properties of carbonAssetsData
-        },
-    },
-  
-  methods: {
-      
-      calculatePercentage(quantity) {
-          // You can calculate the percentage here based on the total quantity
-          if (quantity === 0) {
-              return 0;
-              }
-              const totalQuantity = this.carbonAssetsData.reduce((total, asset) => total + asset.qty, 0);
-              return ((quantity / totalQuantity) * 100).toExponential(2);
-              
-      },
-                renderChart() {
-                    if (this.carbonAssetsData && this.carbonAssetsData.length > 0) {
-                        const ctx = this.$refs.myChart.getContext('2d');
-                        this.chart = new Chart(ctx, {
-                            type: 'doughnut', // Set your chart type
-                            data: this.carbonAssetsData[this.carbonAssetsData.length - 1].chartData,options: {
-                                responsive: true,
-                                maintainAspectRatio: false,
-                                },
-                            });
-                    }
-                },
-  },
-}
-</script>
-
-<style>
-
-</style>
