--- R-REC-NEW/src/components/Navigation.vue
@@ -1,19 +1,0 @@
-<template>
-<nav class="nav">
-    <div class="logo">
-        <img src="@/assets/images/renewvia.webp" alt="renewvia-logo">
-        <a onclick="toggleMenu()" id="menuButton"><i class="fas fa-bars"></i></a>
-    </div>
-    <div class="hide container " id="info">
-        <ul>
-            <li>ABOUT US</li>
-            <li>WHAT ARE R-RECs</li>
-            <li>PRODUCTS</li>
-            <li class="team">OUR TEAM</li>
-            <li class="global">
-                <i class="fas fa-g"><span class="g"></span></i>globalcorp</li>
-        </ul>
-    </div>
-                    
-</nav>
-</template>
